ai-services/requirements.txt
# GuardFlow AI Services - Dependências
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Machine Learning & AI
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.1.0
imageio==2.31.6
scikit-image==0.22.0

# Natural Language Processing
transformers==4.36.0
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1
sentence-transformers==2.2.2

# Data Processing
pandas==2.1.4
numpy==1.24.4
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Database & Cache
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
alembic==1.13.1

# HTTP & API
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0
websockets==12.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0
typer==0.9.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# External APIs
google-cloud-vision==3.4.4
google-cloud-translate==3.11.3
openai==1.3.7
huggingface-hub==0.19.4

# Utilities
tqdm==4.66.1
click==8.1.7
rich==13.7.0
